test(react/confirmactions): refactor confirm actions tests to fit react v19
#231

re #231
